Implements javascript modules
