+++ .prettierignore
@@ -1,3 +1,2 @@
-BookClass.js
 index.js
-ShowandRemoveBooks.js
+BookModule.js
