+++ index.html
@@ -46,5 +46,6 @@
 
     <footer>Copyright by Microverse 2022. All rights reserved</footer>
     <script type="module" src="./index.js"></script>
+
 </body>
 </html>
