+++ modules/BookModule.js
@@ -1,0 +1,10 @@
+export class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
+
+export default {
+  Book,
+};
